pyproject.toml
[tool.poetry]
name = "qa-guru-python-18-api"
version = "0.1.0"
description = ""
authors = ["Yaroslav Slipchishin <yarik.wade@gmail.com>"]
readme = "README.md"
packages = [{include = "qa_guru_python_18_api"}]

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
pytest = "*"
selene = "2.0.0rc9"
allure-pytest = "^2.14.0"
jsonschema = "^4.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
